feat: Initialize sidebar in MediaLibrary for improved user experience on media frame load
