+++ assets/src/js/media-library/views/attachment-browser.js
@@ -1,4 +1,4 @@
-/* global jQuery, _ */
+/* global _ */
 
 /**
  * Internal dependencies
@@ -7,11 +7,9 @@ import MediaLibraryTaxonomyFilter from './filters/media-library-taxonomy-filter'
 import MediaDateRangeFilter from './filters/media-date-range-filter';
 import MediaRetranscode from './filters/media-retranscode';
 
-import { isAPIKeyValid, isUploadPage, isFolderOrgDisabled } from '../utility';
+import { isAPIKeyValid, isUploadPage } from '../utility';
 
 const AttachmentsBrowser = wp?.media?.view?.AttachmentsBrowser;
-
-const $ = jQuery;
 
 /**
  * Attachment Browser with Custom Filters
@@ -72,61 +70,6 @@ export default AttachmentsBrowser?.extend( {
 					} ).render(),
 				);
 			}
-		}
-
-		const hasActiveSortable = this.$el.find( 'ul.ui-sortable:not(.ui-sortable-disabled)' ).length > 0;
-
-		if ( ! isUploadPage() && ! isFolderOrgDisabled() && ! hasActiveSortable ) {
-			/**
-			 * This timeout with the custom event is necessary to ensure that the media frame is fully loaded before dispatching the event.
-			 */
-			setTimeout( () => {
-				$( '.media-frame' ).removeClass( 'hide-menu' );
-
-				if ( window.elementor ) {
-					const visibleContainers = Array.from( document.querySelectorAll( '.supports-drag-drop' ) ).filter(
-						( container ) => getComputedStyle( container ).display !== 'none',
-					);
-
-					const activeContainer = visibleContainers[ visibleContainers.length - 1 ]; // most recently opened visible one
-
-					if ( activeContainer ) {
-						const menu = activeContainer.querySelector( '.media-frame-menu' );
-						if ( menu ) {
-							menu.querySelectorAll( '#rt-transcoder-media-library-root' ).forEach( ( el ) => el.remove() );
-							const div = document.createElement( 'div' );
-							div.id = 'rt-transcoder-media-library-root';
-							if ( menu.firstChild ) {
-								menu.firstChild.appendChild( div );
-							} else {
-								menu.appendChild( div );
-							}
-						}
-					}
-				} else {
-					// Find all visible media frames (same logic as Elementor)
-					const visibleFrames = Array.from( document.querySelectorAll( '.media-frame' ) ).filter(
-						( frame ) => getComputedStyle( frame ).display !== 'none',
-					);
-
-					const activeFrame = visibleFrames[ visibleFrames.length - 1 ]; // most recently opened visible one
-
-					if ( activeFrame ) {
-						const menu = activeFrame.querySelector( '.media-frame-menu .media-menu' );
-						if ( menu ) {
-							// Remove any existing instances
-							menu.querySelectorAll( '#rt-transcoder-media-library-root' ).forEach( ( el ) => el.remove() );
-							// Create and append new div
-							const div = document.createElement( 'div' );
-							div.id = 'rt-transcoder-media-library-root';
-							menu.appendChild( div );
-						}
-					}
-				}
-
-				const event = new CustomEvent( 'media-frame-opened' );
-				document.dispatchEvent( event );
-			}, 50 );
 		}
 	},
